Renamed packages
 `controller` -> `client`, and `sound` -> `audio`.
