+++ android/src/no/ntnu/idi/tdt4240/client/RiskyTurn.java
@@ -1,10 +1,9 @@
-package no.ntnu.idi.tdt4240;
+package no.ntnu.idi.tdt4240.client;
 
 import android.util.Log;
 
 import java.util.Arrays;
 
-import no.ntnu.idi.tdt4240.controller.IRiskyTurn;
 import no.ntnu.idi.tdt4240.model.TurnModel;
 import no.ntnu.idi.tdt4240.model.data.Territory;
 import no.ntnu.idi.tdt4240.util.TerritoryMap;
